Monday, November.... whatever day it is.
Okay so, I had to take out the profile picture stuff. Register is still
basically the same besides that. The only part of the
registration/confirmation stuff that ISN’T working yet is after the
admin has approved the account, I feel like an EMAIL should be sent to
the new user saying “Welcome, click here to get started on RandomTask,”
etc. THIS FUNCTION DOES NOT WORK YET for some reason. I’d love some
assistance on that!